+++ obsidian/labelable-obsidian/esbuild.config.mjs
@@ -1,0 +1,15 @@
+import esbuild from "esbuild";
+
+const prod = process.argv[2] === "production";
+
+esbuild.build({
+  entryPoints: ["main.ts"],
+  bundle: true,
+  external: ["obsidian"],
+  format: "cjs",
+  target: "ES2018",
+  logLevel: "info",
+  sourcemap: prod ? false : "inline",
+  treeShaking: true,
+  outfile: "main.js",
+}).catch(() => process.exit(1));